requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
discord.py==1.7.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.0.1
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.14
tabulate==0.8.9
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
zipp==3.4.1
